+++ demo/site/src/blocks/ColumnsBlock.tsx
@@ -1,6 +1,7 @@
 "use client";
 import { BlocksBlock, PropsWithData, SupportedBlocks, withPreview } from "@comet/cms-site";
 import { ColumnsBlockData, ColumnsContentBlockData } from "@src/blocks.generated";
+import { LinkListBlock } from "@src/blocks/LinkListBlock";
 import styled from "styled-components";
 
 import { DamImageBlock } from "./DamImageBlock";
@@ -13,6 +14,7 @@ const supportedBlocks: SupportedBlocks = {
     richtext: (props) => <RichTextBlock data={props} />,
     headline: (props) => <HeadlineBlock data={props} />,
     image: (props) => <DamImageBlock data={props} aspectRatio="inherit" />,
+    linkList: (props) => <LinkListBlock data={props} />,
 };
 
 const ColumnsContentBlock = withPreview(