+++ packages/admin/blocks-admin/src/iframebridge/IFrameMessage.ts
@@ -67,6 +67,7 @@ export enum AdminMessageType {
     SelectComponent = "SelectComponent",
     HoverComponent = "HoverComponent",
     ContentScope = "ContentScope",
+    GraphQLApiUrl = "GraphQLApiUrl",
 }
 export interface IAdminBlockMessage {
     cometType: AdminMessageType.Block;
@@ -105,9 +106,17 @@ export interface IAdminContentScopeMessage {
     };
 }
 
+export interface IAdminGraphQLApiUrlMessage {
+    cometType: AdminMessageType.GraphQLApiUrl;
+    data: {
+        graphQLApiUrl: string;
+    };
+}
+
 export type AdminMessage =
     | IAdminBlockMessage
     | IAdminSelectComponentMessage
     | IAdminHoverComponentMessage
     | IAdminContentScopeMessage
-    | IAdminShowOnlyVisibleMessage;
+    | IAdminShowOnlyVisibleMessage
+    | IAdminGraphQLApiUrlMessage;