+++ packages/admin/cms-admin/src/generator/future/generateGrid.ts
@@ -481,6 +481,20 @@ export function generateGrid(
 
     const { gridPropsTypeCode, gridPropsParamsCode } = generateGridPropsCode(props);
     const gridToolbarComponentName = `${gqlTypePlural}GridToolbar`;
+    const dataGridRemoteParameters =
+        config.initialSort || config.queryParamsPrefix
+            ? `{${
+                  config.initialSort
+                      ? ` initialSort: [${config.initialSort
+                            .map((item) => {
+                                return `{field: "${item.field}", sort: "${item.sort}"}`;
+                            })
+                            .join(",\n")} ], `
+                      : ""
+              }
+              ${config.queryParamsPrefix ? `queryParamsPrefix: "${config.queryParamsPrefix}",` : ""}
+              }`
+            : "";
 
     const code = `import { gql, useApolloClient, useQuery } from "@apollo/client";
     import {
@@ -617,15 +631,7 @@ export function generateGrid(
     export function ${gqlTypePlural}Grid(${gridPropsParamsCode}): React.ReactElement {
         ${showCrudContextMenuInActionsColumn ? "const client = useApolloClient();" : ""}
         const intl = useIntl();
-        const dataGridProps = { ...useDataGridRemote(${
-            config.initialSort
-                ? `{ initialSort: [${config.initialSort
-                      .map((item) => {
-                          return `{field: "${item.field}", sort: "${item.sort}"}`;
-                      })
-                      .join(",\n")} ] }`
-                : ""
-        }), ...usePersistentColumnState("${gqlTypePlural}Grid") };
+        const dataGridProps = { ...useDataGridRemote(${dataGridRemoteParameters}), ...usePersistentColumnState("${gqlTypePlural}Grid") };
         ${hasScope ? `const { scope } = useContentScope();` : ""}
         ${gridNeedsTheme ? `const theme = useTheme();` : ""}
 
@@ -731,8 +737,8 @@ export function generateGrid(
                                       ? forwardRowAction
                                           ? `{rowAction && rowAction(params)}`
                                           : `
-                                        <IconButton component={StackLink} pageName="edit" payload={params.row.id}>
-                                            <EditIcon color="primary" />
+                                        <IconButton color="primary" component={StackLink} pageName="edit" payload={params.row.id}>
+                                            <EditIcon />
                                         </IconButton>`
                                       : ""
                               }${