+++ packages/site/cms-site/src/iframebridge/IFrameBridge.tsx
@@ -2,7 +2,7 @@
 
 import { createContext, PropsWithChildren, useCallback, useEffect, useRef, useState } from "react";
 import styled from "styled-components";
-import { useDebouncedCallback } from "use-debounce";
+import { useDebounceCallback } from "usehooks-ts";
 
 import { AdminMessage, AdminMessageType, IFrameMessage, IFrameMessageType } from "./IFrameMessage";
 import { PreviewOverlay } from "./PreviewOverlay";
@@ -41,6 +41,7 @@ export interface IFrameBridgeContext {
     sendMessage: (message: IFrameMessage) => void;
     showOutlines: boolean;
     contentScope: unknown;
+    graphQLApiUrl: string | undefined;
     previewElementsData: OverlayElementData[];
     addPreviewElement: (element: PreviewElement) => void;
     removePreviewElement: (element: PreviewElement) => void;
@@ -60,6 +61,7 @@ export const IFrameBridgeContext = createContext<IFrameBridgeContext>({
         //empty
     },
     contentScope: undefined,
+    graphQLApiUrl: undefined,
     previewElementsData: [],
     removePreviewElement: () => {
         // empty
@@ -76,6 +78,7 @@ export const IFrameBridgeProvider = ({ children }: PropsWithChildren) => {
     const [hoveredAdminRoute, setHoveredAdminRoute] = useState<string | undefined>(undefined);
     const [showOutlines, setShowOutlines] = useState<boolean>(false);
     const [contentScope, setContentScope] = useState<unknown>(undefined);
+    const [graphQLApiUrl, setGraphQLApiUrl] = useState<string>("");
     const [previewElements, setPreviewElements] = useState<PreviewElement[]>([]);
     const [previewElementsData, setPreviewElementsData] = useState<OverlayElementData[]>([]);
 
@@ -126,7 +129,7 @@ export const IFrameBridgeProvider = ({ children }: PropsWithChildren) => {
                 recalculatePreviewElementsData();
             });
 
-            mutationObserver.observe(childrenWrapperRef.current, { childList: true, subtree: true });
+            mutationObserver.observe(childrenWrapperRef.current, { childList: true, subtree: true, attributes: true });
             resizeObserver.observe(childrenWrapperRef.current);
 
             return () => {
@@ -140,7 +143,7 @@ export const IFrameBridgeProvider = ({ children }: PropsWithChildren) => {
         window.parent.postMessage(JSON.stringify(message), "*");
     };
 
-    const debounceDeactivateOutlines = useDebouncedCallback(() => {
+    const debounceDeactivateOutlines = useDebounceCallback(() => {
         setShowOutlines(false);
     }, 2500);
 
@@ -167,6 +170,9 @@ export const IFrameBridgeProvider = ({ children }: PropsWithChildren) => {
                     break;
                 case AdminMessageType.ContentScope:
                     setContentScope(message.data.contentScope);
+                    break;
+                case AdminMessageType.GraphQLApiUrl:
+                    setGraphQLApiUrl(message.data.graphQLApiUrl);
                     break;
             }
         },
@@ -232,6 +238,7 @@ export const IFrameBridgeProvider = ({ children }: PropsWithChildren) => {
                 },
                 sendMessage,
                 contentScope,
+                graphQLApiUrl,
                 previewElementsData,
                 addPreviewElement,
                 removePreviewElement,