+++ packages/site/cms-site/src/index.ts
@@ -25,7 +25,14 @@ export { useLocalStorageCookieApi } from "./cookies/useLocalStorageCookieApi";
 export { useOneTrustCookieApi } from "./cookies/useOneTrustCookieApi";
 export { ErrorHandlerBoundary } from "./errorHandler/ErrorHandlerBoundary";
 export { ErrorHandlerProvider } from "./errorHandler/ErrorHandlerProvider";
-export { createFetchWithDefaults, createFetchWithPreviewHeaders, createGraphQLFetch, gql, GraphQLFetch } from "./graphQLFetch/graphQLFetch";
+export {
+    convertPreviewDataToHeaders,
+    createFetchWithDefaults,
+    createFetchWithPreviewHeaders,
+    createGraphQLFetch,
+    gql,
+    GraphQLFetch,
+} from "./graphQLFetch/graphQLFetch";
 export { IFrameBridgeProvider } from "./iframebridge/IFrameBridge";
 export { IFrameMessageType } from "./iframebridge/IFrameMessage";
 export { Preview } from "./iframebridge/Preview";
@@ -37,8 +44,9 @@ export { calculateInheritAspectRatio, generateImageUrl, getMaxDimensionsFromArea
 export { BlockPreviewProvider } from "./preview/BlockPreviewProvider";
 export { usePreview } from "./preview/usePreview";
 export { PreviewSkeleton } from "./previewskeleton/PreviewSkeleton";
+export { previewParams, sitePreviewRoute } from "./sitePreview/appRouter/sitePreviewRoute";
 export { sendSitePreviewIFrameMessage } from "./sitePreview/iframebridge/sendSitePreviewIFrameMessage";
 export { SitePreviewIFrameMessageType } from "./sitePreview/iframebridge/SitePreviewIFrameMessage";
 export { legacyPagesRouterSitePreviewApiHandler } from "./sitePreview/pagesRouter/legacyPagesRouterSitePreviewApiHandler";
 export { SitePreviewProvider } from "./sitePreview/SitePreviewProvider";
-export { previewParams, SitePreviewData, SitePreviewParams, sitePreviewRoute } from "./sitePreview/SitePreviewUtils";
+export { SitePreviewData, SitePreviewParams } from "./sitePreview/SitePreviewUtils";