+++ storybook/src/admin/mui/Link.stories.tsx
@@ -1,0 +1,14 @@
+import { Link } from "@mui/material";
+import * as React from "react";
+
+export default {
+    title: "@comet/admin/mui",
+};
+
+export const _Link = () => {
+    return (
+        <Link href="https://github.com/vivid-planet/comet" target="_blank">
+            Link with Default Styling
+        </Link>
+    );
+};