+++ packages/admin/admin-theme/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/admin-theme
 
+## 7.7.0
+
+### Patch Changes
+
+-   @comet/admin-icons@7.7.0
+
+## 7.6.0
+
+### Patch Changes
+
+-   @comet/admin-icons@7.6.0
+
 ## 7.5.0
 
 ### Patch Changes