+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @comet/eslint-config
 
+## 7.7.0
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.7.0
+
+## 7.6.0
+
+### Patch Changes
+
+-   @comet/eslint-plugin@7.6.0
+
 ## 7.5.0
 
 ### Patch Changes